requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.16.0
pandas==2.1.4
python-dotenv==1.0.0
schedule==1.2.0
mysql-connector-python==8.3.0
